+++ packages/app/src/cli/models/app/app.ts
@@ -7,32 +7,28 @@ import {
   uiExtensions,
   getUIExtensionRendererDependency,
   ExtensionTypes,
+  dotEnvFileNames,
 } from '../../constants'
-import {file, error, path, schema, string, toml, dependency} from '@shopify/cli-kit'
+import {file, error, path, schema, string, toml, dependency, dotenv} from '@shopify/cli-kit'
 
 export const WebConfigurationFileNotFound = (directory: string) => {
   return new error.Abort(`Couldn't find ${configurationFileNames.web} in ${directory}`)
 }
 
 export interface Identifiers {
-  app: {
-    /** API key */
-    apiKey: string
-    /** API secret */
-    apiSecret?: string
-  }
+  /** Application's API Key */
+  app: string
 
   /**
    * The extensions' unique identifiers.
    */
   extensions: {
-    [key: string]: string
+    [directory: string]: string
   }
 }
 
 export const AppConfigurationSchema = schema.define.object({
   name: schema.define.string(),
-  id: schema.define.string().optional(),
   scopes: schema.define.string().default(''),
 })
 
@@ -41,7 +37,6 @@ export type AppConfiguration = schema.define.infer<typeof AppConfigurationSchema
 const UIExtensionConfigurationSchema = schema.define.object({
   name: schema.define.string(),
   type: schema.define.enum(uiExtensions.types),
-  id: schema.define.string().optional(),
   metafields: schema.define
     .array(
       schema.define.object({
@@ -69,22 +64,23 @@ const ThemeExtensionConfigurationSchema = schema.define.object({
 
 type ThemeExtensionConfiguration = schema.define.infer<typeof ThemeExtensionConfigurationSchema>
 
-interface FunctionExtension {
+export interface Extension {
+  idEnvironmentVariableName: string
+  localIdentifier: string
+  configurationPath: string
+  directory: string
+}
+
+export type FunctionExtension = Extension & {
   configuration: FunctionExtensionConfiguration
-  configurationPath: string
-  directory: string
 }
 
-interface ThemeExtension {
+export type ThemeExtension = Extension & {
   configuration: ThemeExtensionConfiguration
-  configurationPath: string
-  directory: string
 }
 
-export interface UIExtension {
+export type UIExtension = Extension & {
   configuration: UIExtensionConfiguration
-  configurationPath: string
-  directory: string
   buildDirectory: string
   entrySourceFilePath: string
 }
@@ -110,19 +106,29 @@ export interface Web {
   configuration: WebConfiguration
 }
 
+export interface AppEnvironment {
+  dotenv: {
+    local?: dotenv.DotEnvFile
+    production?: dotenv.DotEnvFile
+  }
+  env: {[key: string]: string}
+}
+
 export interface App {
+  idEnvironmentVariableName: string
   directory: string
   dependencyManager: dependency.DependencyManager
   configuration: AppConfiguration
   configurationPath: string
   nodeDependencies: {[key: string]: string}
   webs: Web[]
+  environment: AppEnvironment
   extensions: {
     ui: UIExtension[]
     theme: ThemeExtension[]
     function: FunctionExtension[]
   }
-  errors?: string[]
+  errors?: AppErrors
 }
 
 export type AppLoaderMode = 'strict' | 'report'
@@ -132,12 +138,34 @@ interface AppLoaderConstructorArgs {
   mode: AppLoaderMode
 }
 
+class AppErrors {
+  private errors: {
+    [key: string]: string
+  } = {}
+
+  addError(path: string, message: string): void {
+    this.errors[path] = message
+  }
+
+  getError(path: string): string {
+    return this.errors[path]
+  }
+
+  isEmpty() {
+    return Object.keys(this.errors).length === 0
+  }
+
+  toJSON(): string[] {
+    return Object.values(this.errors)
+  }
+}
+
 class AppLoader {
   private directory: string
   private mode: AppLoaderMode
   private appDirectory = ''
   private configurationPath = ''
-  private errors: string[] = []
+  private errors: AppErrors = new AppErrors()
 
   constructor({directory, mode}: AppLoaderConstructorArgs) {
     this.mode = mode
@@ -145,11 +173,11 @@ class AppLoader {
   }
 
   async loaded() {
-    this.errors = []
     this.appDirectory = await this.findAppDirectory()
     const configurationPath = await this.getConfigurationPath()
     const configuration = await this.parseConfigurationFile(AppConfigurationSchema, configurationPath)
     const extensionsPath = path.join(this.appDirectory, `${blocks.extensions.directoryName}`)
+    const environment = await this.loadEnvironment()
     const functions = await this.loadFunctions(extensionsPath)
     const uiExtensions = await this.loadUIExtensions(extensionsPath)
     const themeExtensions = await this.loadThemeExtensions(extensionsPath)
@@ -168,16 +196,45 @@ class AppLoader {
     }
 
     const app: App = {
+      idEnvironmentVariableName: 'SHOPIFY_APP_ID',
       directory: this.appDirectory,
       webs: await this.loadWebs(),
       configuration,
       configurationPath,
+      environment,
       extensions: {ui: uiExtensions, theme: themeExtensions, function: functions},
       dependencyManager,
       nodeDependencies,
     }
-    if (this.errors.length > 0) app.errors = this.errors
+    if (!this.errors.isEmpty()) app.errors = this.errors
     return app
+  }
+
+  async loadEnvironment(systemEnv: {[key: string]: string | undefined} = process.env): Promise<AppEnvironment> {
+    const env = Object.fromEntries(
+      Object.entries(systemEnv).filter(([key, value]) => {
+        return key.startsWith('SHOPIFY_') && value
+      }),
+    ) as {[key: string]: string}
+
+    let localEnv: dotenv.DotEnvFile | undefined
+    let productionEnv: dotenv.DotEnvFile | undefined
+
+    const localEnvPath = path.join(this.appDirectory, dotEnvFileNames.local)
+    if (await file.exists(localEnvPath)) {
+      localEnv = await dotenv.read(localEnvPath)
+    }
+    const productionEnvPath = path.join(this.appDirectory, dotEnvFileNames.production)
+    if (await file.exists(productionEnvPath)) {
+      productionEnv = await dotenv.read(productionEnvPath)
+    }
+    return {
+      dotenv: {
+        production: productionEnv,
+        local: localEnv,
+      },
+      env,
+    }
   }
 
   async findAppDirectory() {
@@ -223,7 +280,7 @@ class AppLoader {
 
   async loadConfigurationFile(path: string): Promise<object> {
     if (!(await file.exists(path))) {
-      return this.abortOrReport(`Couldn't find the configuration file at ${path}`, '')
+      return this.abortOrReport(`Couldn't find the configuration file at ${path}`, '', path)
     }
     const configurationContent = await file.read(path)
     // Convert snake_case keys to camelCase before returning
@@ -240,7 +297,7 @@ class AppLoader {
 
     const parseResult = schema.safeParse(configurationObject)
     if (!parseResult.success) {
-      this.abortOrReport(`Invalid schema in ${path}:\n${JSON.stringify(parseResult.error.issues, null, 2)}`, {})
+      this.abortOrReport(`Invalid schema in ${path}:\n${JSON.stringify(parseResult.error.issues, null, 2)}`, {}, path)
     }
     return parseResult.data
   }
@@ -253,11 +310,13 @@ class AppLoader {
       const directory = path.dirname(configurationPath)
       const configuration = await this.parseConfigurationFile(UIExtensionConfigurationSchema, configurationPath)
       return {
+        idEnvironmentVariableName: `SHOPIFY_${string.constantize(path.basename(directory))}_ID`,
         directory,
         configuration,
         configurationPath,
         buildDirectory: path.join(directory, 'build'),
         entrySourceFilePath: path.join(directory, 'src/index.js'),
+        localIdentifier: path.basename(directory),
       }
     })
     return Promise.all(extensions)
@@ -270,7 +329,13 @@ class AppLoader {
     const functions = configPaths.map(async (configurationPath) => {
       const directory = path.dirname(configurationPath)
       const configuration = await this.parseConfigurationFile(FunctionExtensionConfigurationSchema, configurationPath)
-      return {directory, configuration, configurationPath}
+      return {
+        directory,
+        configuration,
+        configurationPath,
+        idEnvironmentVariableName: `SHOPIFY_${string.constantize(path.basename(directory))}_ID`,
+        localIdentifier: path.basename(directory),
+      }
     })
     return Promise.all(functions)
   }
@@ -282,16 +347,22 @@ class AppLoader {
     const themeExtensions = configPaths.map(async (configurationPath) => {
       const directory = path.dirname(configurationPath)
       const configuration = await this.parseConfigurationFile(ThemeExtensionConfigurationSchema, configurationPath)
-      return {directory, configuration, configurationPath}
+      return {
+        directory,
+        configuration,
+        configurationPath,
+        idEnvironmentVariableName: `SHOPIFY_${string.constantize(path.basename(directory))}_ID`,
+        localIdentifier: path.basename(directory),
+      }
     })
     return Promise.all(themeExtensions)
   }
 
-  abortOrReport(errorMessage: string, fallback: any = null) {
+  abortOrReport(errorMessage: string, fallback: any = null, configurationPath: string) {
     if (this.mode === 'strict') {
       throw new error.Abort(errorMessage)
     } else {
-      this.errors.push(errorMessage)
+      this.errors.addError(configurationPath, errorMessage)
       return fallback
     }
   }
@@ -308,6 +379,88 @@ export async function updateDependencies(app: App): Promise<App> {
   return {
     ...app,
     nodeDependencies,
+  }
+}
+
+export type EnvironmentType = 'local' | 'production'
+
+interface UpdateAppIdentifiersOptions {
+  app: App
+  identifiers: Identifiers
+  environmentType: EnvironmentType
+}
+
+/**
+ * Given an app and a set of identifiers, it persists the identifiers in the .env files.
+ * @param options {UpdateAppIdentifiersOptions} Options.
+ * @returns {App} An copy of the app with the environment updated to reflect the updated identifiers.
+ */
+export async function updateAppIdentifiers({
+  app,
+  identifiers,
+  environmentType,
+}: UpdateAppIdentifiersOptions): Promise<App> {
+  const envVariables = Object.keys(app.environment.env)
+  let dotenvFile = environmentType === 'local' ? app.environment.dotenv.local : app.environment.dotenv.production
+  if (!dotenvFile) {
+    dotenvFile = {
+      path: path.join(app.directory, environmentType === 'local' ? dotEnvFileNames.local : dotEnvFileNames.production),
+      variables: {},
+    }
+  }
+  const variables: {[key: string]: string} = {}
+  if (!envVariables.includes(app.idEnvironmentVariableName)) {
+    variables[app.idEnvironmentVariableName] = identifiers.app
+  }
+  Object.keys(identifiers.extensions).forEach((identifier) => {
+    const envVariable = `SHOPIFY_${string.constantize(identifier)}_ID`
+    if (!envVariables.includes(envVariable)) {
+      variables[envVariable] = identifiers.extensions[identifier]
+    }
+  })
+  dotenvFile.variables = variables
+  await dotenv.write(dotenvFile)
+  return {
+    ...app,
+    environment: {
+      env: app.environment.env,
+      dotenv: {
+        local: environmentType === 'local' ? dotenvFile : app.environment.dotenv?.local,
+        production: environmentType === 'production' ? dotenvFile : app.environment.dotenv?.production,
+      },
+    },
+  }
+}
+
+interface GetAppIdentifiersOptions {
+  app: App
+  environmentType: EnvironmentType
+}
+
+/**
+ * Given an app and a environment, it fetches the ids from the environment
+ * and returns them.
+ * @param options {GetAppIdentifiersOptions} Options.
+ * @returns
+ */
+export function getAppIdentifiers({app, environmentType}: GetAppIdentifiersOptions): Partial<Identifiers> {
+  const envVariables = {
+    ...app.environment.env,
+    ...(environmentType === 'local' ? app.environment.dotenv.local : app.environment.dotenv.production)?.variables,
+  }
+  const extensionsIdentifiers: {[key: string]: string} = {}
+  const processExtension = (extension: Extension) => {
+    if (Object.keys(envVariables).includes(extension.idEnvironmentVariableName)) {
+      extensionsIdentifiers[extension.localIdentifier] = envVariables[extension.idEnvironmentVariableName]
+    }
+  }
+  app.extensions.ui.forEach(processExtension)
+  app.extensions.function.forEach(processExtension)
+  app.extensions.theme.forEach(processExtension)
+
+  return {
+    app: envVariables[app.idEnvironmentVariableName],
+    extensions: extensionsIdentifiers,
   }
 }